+++ flexmeasures/data/models/user.py
@@ -69,6 +69,7 @@ class Account(db.Model, AuthModelMixin):
     secondary_color = Column(String(7), default=None)
     logo_url = Column(String(255), default=None)
     attributes = Column(MutableDict.as_mutable(JSONB), nullable=False, default={})
+    secrets = Column(MutableDict.as_mutable(JSONB), nullable=False, default={})
     annotations = db.relationship(
         "Annotation",
         secondary="annotations_accounts",
@@ -289,6 +290,7 @@ class User(db.Model, UserMixin, AuthModelMixin):
         """
         Within the same account, everyone can read. Consultants as well.
         Only the user themselves, consultants or account-admins can edit their user record.
+        Check policy.can_modify_role() for special treatment of roles.
         Creation and deletion are left to site admins in CLI.
         """
         return {