+++ documentation/cli/change_log.rst
@@ -4,6 +4,12 @@
 FlexMeasures CLI Changelog
 **********************
 
+since v1.0.0 | July XX, 2026
+=================================
+
+* Add ``flexmeasures edit secret`` to store an encrypted secret on an account or asset.
+* Add ``flexmeasures delete secret`` to remove an encrypted secret from an account or asset.
+
 since v0.33.0 | June 01, 2026
 =================================